// types/inventoryVoucher/PurchaseVoucherTypes.d.ts
export interface PurchaseVoucherTableProps {
  loading: boolean;
  handleShowDeleteDialog: (id: number) => void;
  showDeleteDialog: boolean;
  setShowDeleteDialog: Dispatch<SetStateAction<boolean>>;
  setTempDeleteId: Dispatch<SetStateAction<number | null>>;
  handleDeletePurchase: () => void;
  deletePurchaseVoucherLoading: boolean;
  currentPage: number;
  setCurrentPage: Dispatch<SetStateAction<number>>;
  lastPage: number;
}
// ...
export interface PurchaseVoucherTableData {
  Id: number;
  Purchase_Date: string;
  Purchase_No: string;
  Party_Name: string;
  Party_Id: number;
  Total_Amount: string;
  Purchase_Type?: string | number;
  Pur_Type?: string | number;
  Is_Order?: string | number;
  ItemRow: {
    Item_Id: number;
    Item_Name: string;
    Item_Qnty: string;
    Item_Rate: string;
    Item_CGST: string;
    Item_SGST: string;
    Item_IGST: string | null;
  }[];
}